// src/interfaces/crypto.ts
export interface CryptoProps {
    id: number;
    name: string;
    symbol: string;
    slug: string;
    num_market_pairs: number;
    date_added: string;
    tags: string[];
    max_supply: number;
    circulating_supply: number;
    total_supply: number;
    infinite_supply: boolean;
    platform: string | null;
    cmc_rank: 1;
    self_reported_circulating_supply: string | null;
    self_reported_market_cap: string | null;
    tvl_ratio: string | null;
    last_updated: string;
    quote: {
        USD: {
            price: number;
            volume_24h: number;
            volume_change_24h: number;
            percent_change_1h: number;
            percent_change_24h: number;
            percent_change_7d: number;
            percent_change_30d: number;
            percent_change_60d: number;
            percent_change_90d: number;
            market_cap: number;
            market_cap_dominance: number;
            fully_diluted_market_cap: number;
            tvl: string | null;
            last_updated: string
        }
    }
}